GPos objects now exist in 2 flavors: UnstitchedGPos and StitchedGPos
GPos is now a virtual class with 2 concrete subclasses: UnstitchedGPos
and StitchedGPos. In an UnstitchedGPos instance the positions are stored
as an integer vector. In a StitchedGPos instance, like with old GPos
instances, the positions are stored as an IRanges object where each range
represents a run of consecutive positions. This is analog to the
IPos/UnstitchedIPos/StitchedIPos situation. See ?GPos for more information.

Old serialized GPos instances will need to be converted to StitchedGPos
instances with updateObject() (not ready yet).